% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGCR.R
\name{BGCR}
\alias{BGCR}
\title{Bayesian Graphical compositional Regression}
\usage{
BGCR(PrJAP = 0.5, sum_PrMAP = "default", threshold = 0.005,
  kappa = 0, tree, otu_group_1, otu_group_2, X_group_1 = "default",
  X_group_2 = "default", nu = 10^(seq(-1, 4)), sigma = 4,
  verbose = FALSE)
}
\arguments{
\item{PrJAP}{a value between 0 and 1. The prior joint alternative probability used for choosing \eqn{\alpha}.}

\item{sum_PrMAP}{a positive value less than the total number of internal nodes in the phylogenetic tree. The sum of prior marginal alternative probability used for choosing \eqn{\tau}. The default value is set to 2, recommended for 100 OTUs.}

\item{threshold}{a positive value controls the accuracy used for the empirical Bayes estimate of \eqn{\tau}.}

\item{kappa}{a non-negative value controls the chaining pattern will occur along the phylogenetic tree. The default value 0 introduces an explaning away effect.}

\item{tree}{an object of class ``phylo'' defined by the \code{ape} package that summarizes the phylogenetic information of the OTUs.}

\item{otu_group_1}{a matrix containing the OTU samples of the first group. Each row represents an OTU, each column represents a sample.}

\item{otu_group_2}{a matrix containing the OTU samples of the second group, same as \code{otu_group_1}.}

\item{X_group_1}{a matrix containg the covariate of samples in the first group that need to be adjusted in the test. Each row represetns a sample, each column represents a covariate. The default setting adjusts for no covariate.}

\item{X_group_2}{a matrix containg the covariate of samples in the second group, same as \code{X_group_2}.}

\item{nu}{a vector contains a sequence of the dispersion parameter \eqn{\nu} used for numerically computing the marginal likelihood.}

\item{sigma}{a positive value. The standard deviation of the normal prior on the regression coefficient.}

\item{verbose}{logicals. If true, print the number of internal node processed while computing the marginal likelihoods.}
}
\value{
\code{BGCR} returns an object of class \code{"BGCR"}. A \code{"BGCR"} object is a list containing the following components:
 \itemize{
  \item \code{tree} - the phylogenetic tree. An object of class \code{"phylo"} containing the phylogenetic information of the OTUs.
  \item \code{PJAP} - the posterior joint alternative probability.
  \item \code{PMAP} - a vector containing the posterior marginal alternative probability at each internal node of the phylogenetic tree, ordered according to the label of the node in the phylogenetic tree. Note that the first half of the vector contains zeros since they represent the leaves of the tree.
  \item \code{PJAP_ind} - the posterior joint alternative probability returned by BCR, similar to \code{PJAP}.
  \item \code{PMAP_ind} - the posterior marginal alternative probabilities returned by BCR, similar to \code{PMAP}.
  \item \code{BF} - a vector containing the Bayes factor comparing the local hypotheses at each internal node. The orders are same as the orders in \code{PMAP}.
  \item \code{alpha} - the value of \eqn{\alpha} used for BGCR.
  \item \code{tau} - the value of \eqn{\tau} used for BGCR estimated by the empirical Bayes procedure.
}
}
\description{
Fit the Bayesian Graphical Compositional Regression for comparing two groups of microbiome count data.
}
\examples{
library(ape)
exp_tree <- rtree(4) #create a phylogenetic tree with 4 leaves

#generate the OTU table for the first group
otu1 <- matrix(c(2,3,4,6,3,5,5,8), nrow = 4, byrow = TRUE)

#generate the OTU table for the second group
otu2 <- matrix(c(1,1,1,1,2,2,2,2), nrow = 4, byrow = TRUE)

rownames(otu1) <- exp_tree$tip.label #name the OTUs in the OTU table
rownames(otu2) <- exp_tree$tip.label #name the OTUs in the OTU table

result <- BGCR(PrJAP = 0.5, sum_PrMAP = "default", threshold = 0.005,
               kappa = 0, tree = exp_tree, otu_group_1 = otu1, otu_group_2 = otu2)

plot(x = result)

}
